# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Microservices FTW is a sample Swagger\n                  project using Jetty and Jersey.",
    "version" : "v1",
    "title" : "Microservices FTW REST API",
    "termsOfService" : "http://terms-and-conditions.com",
    "contact" : {
      "name" : "Nelson Gomes",
      "url" : "http://www.bitsighttech.com",
      "email" : "nelson.ricardo.gomes@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath" : "/api",
  "schemes" : [ "http" ],
  "paths" : {
    "/v1/books" : {
      "get" : {
        "summary" : "List Books method",
        "description" : "Method to get a books list.",
        "operationId" : "getBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Book"
              }
            }
          },
          "500" : {
            "description" : "unsuccessful operation"
          }
        }
      },
      "post" : {
        "summary" : "Create Books method",
        "description" : "Method to create one or more books.",
        "operationId" : "createBooks",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Json array of Books",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/Book"
            }
          }
        }, {
          "name" : "comment",
          "in" : "query",
          "description" : "sample comment",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "500" : {
            "description" : "unsuccessful operation"
          }
        }
      }
    },
    "/v1/counter" : {
      "get" : {
        "summary" : "Get your cluster request counter",
        "description" : "Method to get a cluster counter.",
        "operationId" : "getCounter",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      }
    },
    "/v1/ping" : {
      "get" : {
        "summary" : "Ping method",
        "description" : "Method to check API connectivity.",
        "operationId" : "pingMethod",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "Book" : {
      "type" : "object",
      "required" : [ "isbn", "price", "publisher", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32",
          "example" : "1",
          "description" : "Book id"
        },
        "title" : {
          "type" : "string",
          "example" : "Book Title",
          "description" : "Book title"
        },
        "isbn" : {
          "type" : "string",
          "example" : "Book ISNTitle",
          "description" : "Book ISBN"
        },
        "price" : {
          "type" : "number",
          "format" : "float",
          "example" : "12.12",
          "description" : "Book Price"
        },
        "publisher" : {
          "type" : "string",
          "example" : "Publisher",
          "description" : "Publisher Name"
        }
      }
    }
  }
}